requirements.txt
discord.py>=2.4.0
colorama==0.4.6
aiohttp>=3.7.4